Updated /core and /daq, removed trout_wizard from build
- Updated /daq to version (3.0.27):

commit 959b13d76aa3409acadfe22d1c30864698a297a7 (HEAD, tag: v3.0.27, origin/master, origin/HEAD, master)
Author: Priyanka Gurudev (prbg) <prbg@cisco.com>
Date:   Wed Mar 18 01:33:32 2026 +0000

- Updated /core to version (3.12.2.0):

commit 14aeb09f5a0856812dbe08ead3c21f99e8860aa0 (tag: 3.12.2.0, upstream/master, origin/master, origin/HEAD, master)
Author: Priyanka Gurudev <prbg@cisco.com>
Date:   Thu Apr 23 14:38:57 2026 -0400

- Removed trout_wizard from build
Updating to version 3.12.x.x of snort3 removes the dedicated support for wizards, they are now just specialized service inspectors, as the current wizard was just an attempt it has not been converted, but the code has been removed from the build instead.

- Updated dhcp_option due to minor changes in the interface of snort3

+++ core/src/stream/tcp/tcp_segment_descriptor.h
@@ -1,5 +1,5 @@
 //--------------------------------------------------------------------------
-// Copyright (C) 2015-2025 Cisco and/or its affiliates. All rights reserved.
+// Copyright (C) 2015-2026 Cisco and/or its affiliates. All rights reserved.
 //
 // This program is free software; you can redistribute it and/or modify it
 // under the terms of the GNU General Public License Version 2 as published
@@ -128,10 +128,10 @@ public:
         pkt->dsize -= offset;
     }
 
-    void set_packet_flags(uint32_t flags) const
+    void set_packet_flags(uint64_t flags) const
     { pkt->packet_flags |= flags; }
 
-    bool are_packet_flags_set(uint32_t flags) const
+    bool are_packet_flags_set(uint64_t flags) const
     { return (pkt->packet_flags & flags) == flags; }
 
     uint32_t get_packet_timestamp() const
